requirements.txt
flask
pandas
openai
python-dotenv